Fix test after merge. Move EvaluateLogger to a new file.
