+++ src/commonTest/kotlin/com/configcat/ConfigServiceTests.kt
@@ -426,9 +426,9 @@ class ConfigServiceTests {
     @Test
     fun testCacheTTLRespectsExternalCache() = runTest {
         val mockEngine = MockEngine {
-            respond(content = Data.formatJsonBody("test_remote"), status = HttpStatusCode.OK)
+            respond(content = Data.formatJsonBodyWithString("test_remote"), status = HttpStatusCode.OK)
         }
-        val cache = SingleValueCache(Data.formatCacheEntryWithEtag("test_local", "etag"))
+        val cache = SingleValueCache(Data.formatCacheEntryWithETag("test_local", "etag"))
         val service = Services.createConfigService(
             mockEngine,
             lazyLoad {
@@ -436,12 +436,12 @@ class ConfigServiceTests {
             },
             cache
         )
-        assertEquals("test_local", service.getSettings().settings["fakeKey"]?.value.toString())
+        assertEquals("test_local", service.getSettings().settings["fakeKey"]?.settingsValue?.stringValue)
         assertEquals(0, mockEngine.requestHistory.size)
         TestUtils.wait(1.seconds)
 
-        cache.write("", Data.formatCacheEntryWithEtag("test_local2", "etag2"))
-        assertEquals("test_local2", service.getSettings().settings["fakeKey"]?.value.toString())
+        cache.write("", Data.formatCacheEntryWithETag("test_local2", "etag2"))
+        assertEquals("test_local2", service.getSettings().settings["fakeKey"]?.settingsValue?.stringValue)
         assertEquals(0, mockEngine.requestHistory.size)
     }
